Rename to FakePlatformChecker and move to own file
