+++ dart/test/fake_platform_checker.dart
@@ -1,0 +1,40 @@
+import 'package:sentry/src/platform_checker.dart';
+
+class FakePlatformChecker implements PlatformChecker {
+  FakePlatformChecker.releaseMode() {
+    _releaseMode = true;
+    _debugMode = false;
+    _profileMode = false;
+  }
+
+  FakePlatformChecker.debugMode() {
+    _releaseMode = false;
+    _debugMode = true;
+    _profileMode = false;
+  }
+
+  FakePlatformChecker.profileMode() {
+    _releaseMode = false;
+    _debugMode = false;
+    _profileMode = true;
+  }
+
+  bool _releaseMode;
+  bool _debugMode;
+  bool _profileMode;
+
+  @override
+  bool isReleaseMode() {
+    return _releaseMode;
+  }
+
+  @override
+  bool isDebugMode() {
+    return _debugMode;
+  }
+
+  @override
+  bool isProfileMode() {
+    return _profileMode;
+  }
+}
